fix: Dark mode styles for document diffs
