+++ shared/i18n/locales/en_US/translation.json
@@ -64,6 +64,7 @@
   "Collection": "Collection",
   "Debug": "Debug",
   "Document": "Document",
+  "Revision": "Revision",
   "Navigation": "Navigation",
   "People": "People",
   "Recent searches": "Recent searches",
